ml_service/requirements.txt
flask
flask-cors
opencv-python
numpy
ultralytics
torch
requests
redis
shapely
psycopg2-binary
SQLAlchemy